Fixed the problem in finding the home location for the morning trips
1. Previously the time checking was ">school_start" which is set as 7.
So the time needs to be 8 to meet the time requirement. But in the
trips, the trip time was between 7 to 8, so it was not caught. So ">"
was changed to ">=.

2. To avoid generating too many pois with similar location, the getting
poi part was moved to be called when a trip is detected. Also, when
getting the location of the poi, instead of getting the last location of
this segment, the most frequent location is picked up, because situation
happend where there are many lat values very close to each other but not
completely same.